+++ src/packages/piral-blazor/src/internal/globals.ts
@@ -1,0 +1,5 @@
+import '@dotnet/jsinterop';
+
+window['Blazor'] = {
+  _internal: {},
+};