+++ src/packages/piral-cli/src/common/index.ts
@@ -1,7 +1,10 @@
+export * from './archive';
 export * from './bundler-patches';
 export * from './bundler';
 export * from './compatibility';
 export * from './config';
+export * from './declaration';
+export * from './emulator';
 export * from './emoji';
 export * from './envs';
 export * from './hash';
@@ -21,4 +24,3 @@ export * from './scaffold';
 export * from './scripts';
 export * from './settings';
 export * from './template';
-export * from './untar';