+++ src/packages/piral-cli/src/common/log.ts
@@ -1,10 +1,39 @@
 import * as logger from '@parcel/logger';
 import * as messages from '../messages';
+import { join } from 'path';
 import { format } from 'util';
+import { createWriteStream } from 'fs';
+import { isWindows } from './info';
 import { LogLevels, QuickMessage } from '../types';
 
 type Messages = typeof messages;
 type MessageTypes = keyof Messages;
+
+// unfortunately, Parcel's support for verbose logging on Windows is broken
+if (isWindows) {
+  const stripAnsi = require('strip-ansi');
+
+  logger.prototype.verbose = function(message: string) {
+    if (this.logLevel < 4) {
+      return;
+    }
+
+    let currDate = new Date();
+    message = `[${currDate.toLocaleTimeString()}]: ${message}`;
+    if (this.logLevel > 4) {
+      if (!this.logFile) {
+        // the critical line is the filename; it must not contain colons!
+        const timestamp = currDate.toISOString().replace(/:/g, '');
+        const fileName = `parcel-debug-${timestamp}.log`;
+        this.logFile = createWriteStream(join(process.cwd(), fileName));
+      }
+
+      this.logFile.write(stripAnsi(message) + '\n');
+    }
+
+    this._log(message);
+  };
+}
 
 export function setLogLevel(logLevel: LogLevels) {
   logger.setOptions({ logLevel });